refactor: CreateComment 컴포넌트 CommentList로 merge
