+++ client/src/components/postdetail/commentType.ts
@@ -1,0 +1,10 @@
+export interface CommentType {
+   boardId: number;
+   commentId: number;
+   nickname: string;
+   userPhoto: string;
+   content: string;
+   createdAt: string;
+   parentId: number;
+   replies?: CommentType[];
+}
